fix clang conversion error

+++ rostime/src/time.cpp
@@ -251,7 +251,7 @@ namespace ros
       }
     return 0;
 #else
-    timespec req = { sec, nsec };
+    timespec req = { (long)sec, (long)nsec };
     return nanosleep(&req, NULL);
 #endif
   }
@@ -266,7 +266,7 @@ namespace ros
 #if defined(WIN32)
     ros_nanosleep(sec,nsec);
 #else
-    timespec req = { sec, nsec };
+    timespec req = { (long)sec, (long)nsec };
     timespec rem = {0, 0};
     while (nanosleep(&req, &rem) && !g_stopped)
       {
